Refine comments and tweak code in pkg/scheduler for clarity and readability
